Move user friends functions to userFriends.ts
